+++ src/main/java/com/diginamic/BestiolesREST/service/PersonService.java
@@ -1,11 +1,15 @@
 package com.diginamic.BestiolesREST.service;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.diginamic.BestiolesREST.entity.Person;
+import com.diginamic.BestiolesREST.exception.EntityToCreateHasAnIdException;
+import com.diginamic.BestiolesREST.exception.EntityToUpdateHasNoIdException;
 import com.diginamic.BestiolesREST.repository.PersonRepository;
 
 import jakarta.persistence.EntityNotFoundException;
@@ -18,10 +22,20 @@ public class PersonService {
 	private PersonRepository personRepository;
 
 	public Person create(@Valid Person personToCreate) {
+		if(personToCreate.getId() != null) {
+			throw new EntityToCreateHasAnIdException("Entity To Create Has An Id");
+		}
 		return this.personRepository.save(personToCreate);
 	}
 	
 	public Person update(@Valid Person updatedPerson) {
+		if(updatedPerson.getId() == null) {
+			throw new EntityToUpdateHasNoIdException("Entity To Update Has No Id");
+		}
+		Optional<Person> personOpt = personRepository.findById(updatedPerson.getId());
+		if(personOpt.isEmpty()) {
+			throw new EntityNotFoundException("Entity Not Found");
+		}
 		return this.personRepository.save(updatedPerson);
 	}
 	
@@ -30,10 +44,14 @@ public class PersonService {
 	}
 	
 	public Person findById(Integer id) {
-		return this.personRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+		return this.personRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Entity Not Found"));
 	}
 	
 	public void deleteById(Integer id) {
+		Optional<Person> personOpt = personRepository.findById(id);
+		if(personOpt.isEmpty()) {
+			throw new EntityNotFoundException("Entity Not Found");
+		}
 		personRepository.deleteById(id);
 	}
